added stylecop linter and mods for warnings
