+++ hcgraph.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 25.0.1704.2
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "hcgraph", "hcgraph\hcgraph.csproj", "{78BB7D20-22F3-454A-84D5-4939A5B226AF}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HcGraph", "HcGraph\HcGraph.csproj", "{78BB7D20-22F3-454A-84D5-4939A5B226AF}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
