+++ hcgraph/Domain/Repositories/IItemRepository.cs
@@ -1,0 +1,13 @@
+using HcGraph.Domain.Models;
+
+namespace HcGraph.Domain.Repositories
+{
+    public interface IItemRepository
+    {
+        public Task<List<Item>> GetItems();
+
+        public Task<Item?> GetItem(long rowID);
+        
+        public Task<Item?> GetItemByItemNumber(string itemNumber);
+    }
+}
