+++ hcgraph/Domain/Repositories/IOrderRepository.cs
@@ -1,0 +1,11 @@
+using HcGraph.Domain.Models;
+
+namespace HcGraph.Domain.Repositories
+{
+    public interface IOrderRepository
+    {
+        public Task<List<Order>> GetOrders();
+
+        public Task<Order?> GetOrder(long rowId);
+    }
+}
