+++ hcgraph/Domain/Repositories/OrderItemRepository.cs
@@ -1,16 +1,8 @@
-﻿using hcgraph.Domain.Models;
+﻿using HcGraph.Domain.Models;
 using Microsoft.EntityFrameworkCore;
 
-namespace hcgraph.Domain.Repositories
+namespace HcGraph.Domain.Repositories
 {
-    public interface IOrderItemRepository
-    {
-        public void Create(OrderItem orderItem);
-        public Task<List<OrderItem>> GetOrderItems();
-        public Task<OrderItem?> GetOrderItem(long rowId);
-        public Task<List<OrderItem>> GetOrderItemsByOrderId(long orderId);
-    }
-
     public class OrderItemRepository : IOrderItemRepository
     {
         private readonly SampleDbContext _dbContext;
@@ -21,7 +13,9 @@ namespace hcgraph.Domain.Repositories
         }
 
         public Task<List<OrderItem>> GetOrderItems() => _dbContext.OrderItems.ToListAsync();
+
         public Task<OrderItem?> GetOrderItem(long rowId) => _dbContext.OrderItems.Where(oi => oi.RowId == rowId).FirstOrDefaultAsync();
+
         public Task<List<OrderItem>> GetOrderItemsByOrderId(long orderId) => _dbContext.OrderItems.Where(oi => oi.OrderId == orderId).ToListAsync();
 
         public void Create(OrderItem orderItem)
@@ -31,4 +25,3 @@ namespace hcgraph.Domain.Repositories
         }
     }
 }
-
