+++ hcgraph/Domain/Repositories/OrderRepository.cs
@@ -1,15 +1,8 @@
-﻿using System;
-using hcgraph.Domain.Models;
+﻿using HcGraph.Domain.Models;
 using Microsoft.EntityFrameworkCore;
 
-namespace hcgraph.Domain.Repositories
+namespace HcGraph.Domain.Repositories
 {
-    public interface IOrderRepository
-    {
-        public Task<List<Order>> GetOrders();
-        public Task<Order?> GetOrder(long rowId);
-    }
-
     public class OrderRepository : IOrderRepository
     {
         private readonly SampleDbContext _dbContext;
@@ -20,7 +13,7 @@ namespace hcgraph.Domain.Repositories
         }
 
         public Task<List<Order>> GetOrders() => _dbContext.Orders.ToListAsync();
+
         public Task<Order?> GetOrder(long rowId) => _dbContext.Orders.Where(o => o.RowId == rowId).FirstOrDefaultAsync();
     }
 }
-
