+++ hcgraph/Queries/ItemQuery.cs
@@ -1,13 +1,13 @@
-﻿using hcgraph.Domain.Models;
-using hcgraph.Domain.Services;
+﻿using HcGraph.Domain.Models;
+using HcGraph.Domain.Services;
 
-namespace hcgraph.Queries
+namespace HcGraph.Queries
 {
     [ExtendObjectType("Query")]
     public class ItemQuery
     {
         public async Task<List<Item>> GetItems([Service] IOrderService orderService) => await orderService.GetItems();
+        
         public async Task<Item?> GetItem([Service] IOrderService orderService, long rowId) => await orderService.GetItem(rowId);
     }
 }
-
